Cleaning up the tests
We keep having cascades of test failures if tests 1-3 fail. It looks
like a whole bunch of stuff fails, which is essentially just a bunch of
false-positives.

While it is likely that one of those testing indicates a full failure
(usually one of connectivity), it doesn’t necessarily mean it. This
will allow us to pass more tests if there are minor connectivity
hiccups, and maybe better lead us to the real cause of actual library
bugs.