DESTROY: Only close connection from opener PID
This fixes: REQUEST: optionally leave connection alive in
net_amqp_rabbitmq_DESTROY to allow forking #151

The problem was with forks. If an AMQP connection is open when the fork
occurs both the parent and the child process would close the
connection in DESTROY. It should only be closed by the same process
that opened it.

$self is really a amqp_connection_state_t defined in
rabbitmq-include/amqp.h - one of RabbitMQ's header files. So there is
nowhere to store the creator's PID.

To remedy that, new and DESTROY have been turned into perl subs, that
call their renamed XS counterparts. new() stores the PID in a
fieldhash using Hash::Util::FieldHash inside-out object (see perlobj)
after calling _new(), and DESTROY checks to see whether to close the
connection by comparing the current PID with the PID that created the
connection in new().

Added Hash::Util::FieldHash as a dependency - in core since v5.9.4