Use content_type => 'binary' instead of 'C'.
The existing code used:

    content_type => 'C'

and in some places:

    content_type => 'none'

'C' looks like it came about as a parallel to LANG=C. But MIME is not
locales.

RFC-2045:
                 Multipurpose Internet Mail Extensions
                            (MIME) Part One:
                   Format of Internet Message Bodies

which obsoletes RFC-1341 MIME (Multipurpose Internet Mail Extensions),
defines:

> 6.1.  Content-Transfer-Encoding Syntax
>
>    The Content-Transfer-Encoding field's value is a single token
>    specifying the type of encoding, as enumerated below.  Formally:
>
>      encoding := "Content-Transfer-Encoding" ":" mechanism
>
>      mechanism := "7bit" / "8bit" / "binary" /
>                   "quoted-printable" / "base64" /
>                   ietf-token / x-token
>
>    These values are not case sensitive -- Base64 and BASE64 and bAsE64
>    are all equivalent.  An encoding type of 7BIT requires that the body
>    is already in a 7bit mail-ready representation.  This is the default
>    value -- that is, "Content-Transfer-Encoding: 7BIT" is assumed if the
>    Content-Transfer-Encoding header field is not present.

And the RabbitMQ documentation https://www.rabbitmq.com/consumers.html
's "Content Type and Encoding" section implies that the
content_encoding field is a parallel to the MIME field:

>  The content (MIME media) type and content encoding fields allow
>  publishers communicate how message payload should be deserialized
>  and decoded by consumers.

So we use:

    content_type => 'binary'

when we mean binary to be in line with the RFC-2045. Especially
because users will look in t/* to see how to do things.